+++ legacy_minpack/hybrj1_.c
