+++ legacy_minpack/lmstr1_.c
