+++ legacy_minpack/lmstr_.c
