+++ src/chkder.c
