+++ src/enorm.c
