+++ src/hybrj.c
