+++ src/hybrj1.c
