+++ src/lmstr.c
