+++ src/lmstr1.c
