+++ src/qrfac.c
