+++ src/qrsolv.c
